Duplication of environment prior to work on the harder builtins; relocation of some auxiliary functions to a separate folder
